backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
pydantic==2.6.4
pydantic-settings==2.2.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
slowapi==0.1.9
requests==2.31.0
qrcode==7.4.2
pydantic[email]
numpy==1.26.4
scikit-learn==1.8.0
joblib==1.5.3
shap==0.45.1
pandas==3.0.1
seaborn==0.13.2
flower==2.0.1
imbalanced-learn==0.14.1
scipy==1.12.0
matplotlib==3.8.3
reportlab==4.2.2
redis==5.0.7
fastapi-limiter==0.1.6
rq==1.16.2
python-multipart==0.0.22
torch==2.2.2
torchvision==0.17.2
pillow==10.3.0
opencv-python==4.9.0.80